+++ NOLF2/ObjectDLL/ObjectShared/AIGoalAbstractSearch.cpp
@@ -53,7 +53,7 @@ void CAIGoalAbstractSearch::Save(ILTMessage_Write *pMsg)
 
 	HOBJECT hRegion;
 	AISEARCHREGION_LIST::iterator it;
-	SAVE_DWORD( m_lstRegions.size() );
+	SAVE_DWORD( (uint32)(m_lstRegions.size()) );
 	for( it = m_lstRegions.begin(); it != m_lstRegions.end(); ++it )
 	{
 		hRegion = *it;
