+++ NOLF2/ObjectDLL/ObjectShared/AINodeMgr.cpp
@@ -160,7 +160,7 @@ void CAINodeMgr::Save(ILTMessage_Write *pMsg)
 {
 	SAVE_BOOL( m_bInitialized );
 
-	SAVE_DWORD( m_mapAINodes.size() );
+	SAVE_DWORD( (uint32)(m_mapAINodes.size()) );
 
 	EnumAINodeType eNodeType;
 	AINode* pNode;
@@ -704,7 +704,7 @@ AINode* CAINodeMgr::FindRandomNodeFromThreat(CAI* pAI, EnumAINodeType eNodeType,
 
 	if( !s_lstTempNodes.empty() )
 	{
-		pNode = s_lstTempNodes[ GetRandom( 0, s_lstTempNodes.size() - 1 ) ];
+		pNode = s_lstTempNodes[ GetRandom( 0, (int)(s_lstTempNodes.size()) - 1 ) ];
 		s_lstTempNodes.clear();
 	}
 	else {
@@ -1188,7 +1188,7 @@ AINode* CAINodeMgr::FindRandomOwnedNode(CAI* pAI, EnumAINodeType eNodeType, HOBJ
 
 	if( !s_lstTempNodes.empty() )
 	{
-		pNode = s_lstTempNodes[ GetRandom( 0, s_lstTempNodes.size() - 1 ) ];
+		pNode = s_lstTempNodes[ GetRandom( 0, (int)(s_lstTempNodes.size()) - 1 ) ];
 		s_lstTempNodes.clear();
 	}
 	else {
