+++ NOLF2/ObjectDLL/ObjectShared/AIRegion.cpp
@@ -189,7 +189,7 @@ void AIRegion::Save(ILTMessage_Write *pMsg)
 	SAVE_DWORD(m_cSearchers);
 	SAVE_DWORD(m_cPostSearchMsgs);
 	
-	SAVE_DWORD(m_lstVolumes.size());
+	SAVE_DWORD((uint32)(m_lstVolumes.size()));
 
 	AIVolume* pVolume;
 	AIREGION_VOLUME_LIST::iterator it;
